Removed some fishy checks on PWM values. Now either control duty cycle with 12bit value, or set the channel as GPIO using dedicated method. No magic under the hood, all plain and straight-forward. Also, changed motors disengagement by setting all PCAs' channels to zero with dedicated ALL_LED_XXX registers.

+++ src/ContinuousServo.h
@@ -77,7 +77,7 @@ private:
 	/** A PCA9685 channel to which this servo is connected. */
 	uint8_t mChannel;
 	/** A minimum PWM duty for steering control. */
-	uint16_t mMinDuty;
+	float mMinDuty;
 	/** Duty range for the whole steering range. */
 	float mDutyRange;
 };
